requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
geoalchemy2==0.14.2
psycopg2-binary>=2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-decouple==3.8

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Geospatial
shapely==2.0.2
geopy==2.4.0

# Caching & Async
redis==5.0.1
celery==5.3.4
aioredis==2.0.1

# AI/LLM Integration
openai==1.3.5
langchain==0.0.350

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
